Make classes from samples a top-level class
closes #1126
