+++ spring-boot-admin-samples/spring-boot-admin-sample-servlet/src/main/java/de/codecentric/boot/admin/SpringBootAdminServletApplication.java
@@ -16,101 +16,31 @@
 
 package de.codecentric.boot.admin;
 
-import de.codecentric.boot.admin.server.config.AdminServerProperties;
 import de.codecentric.boot.admin.server.config.EnableAdminServer;
 import de.codecentric.boot.admin.server.domain.entities.InstanceRepository;
-import de.codecentric.boot.admin.server.notify.CompositeNotifier;
-import de.codecentric.boot.admin.server.notify.Notifier;
-import de.codecentric.boot.admin.server.notify.RemindingNotifier;
-import de.codecentric.boot.admin.server.notify.filter.FilteringNotifier;
 import de.codecentric.boot.admin.server.web.client.HttpHeadersProvider;
 import de.codecentric.boot.admin.server.web.client.InstanceExchangeFilterFunction;
 
-import java.time.Duration;
-import java.util.Collections;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Profile;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
-import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
 
 @Configuration
 @EnableAutoConfiguration
 @EnableAdminServer
+@Import({SecurityPermitAllConfig.class, SecuritySecureConfig.class, NotifierConfig.class})
 public class SpringBootAdminServletApplication {
     private static final Logger log = LoggerFactory.getLogger(SpringBootAdminServletApplication.class);
 
     public static void main(String[] args) {
         SpringApplication.run(SpringBootAdminServletApplication.class, args);
     }
-
-    @Profile("insecure")
-    @Configuration
-    public static class SecurityPermitAllConfig extends WebSecurityConfigurerAdapter {
-        private final AdminServerProperties adminServer;
-
-        public SecurityPermitAllConfig(AdminServerProperties adminServer) {
-            this.adminServer = adminServer;
-        }
-
-        @Override
-        protected void configure(HttpSecurity http) throws Exception {
-            http.authorizeRequests()
-                .anyRequest()
-                .permitAll()
-                .and()
-                .csrf()
-                .csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
-                .ignoringAntMatchers(this.adminServer.path("/instances"), this.adminServer.path("/actuator/**"));
-        }
-    }
-
-    @Profile("secure")
-    // tag::configuration-spring-security[]
-    @Configuration
-    public static class SecuritySecureConfig extends WebSecurityConfigurerAdapter {
-        private final AdminServerProperties adminServer;
-
-        public SecuritySecureConfig(AdminServerProperties adminServer) {
-            this.adminServer = adminServer;
-        }
-
-        @Override
-        protected void configure(HttpSecurity http) throws Exception {
-            // @formatter:off
-            SavedRequestAwareAuthenticationSuccessHandler successHandler = new SavedRequestAwareAuthenticationSuccessHandler();
-            successHandler.setTargetUrlParameter("redirectTo");
-            successHandler.setDefaultTargetUrl(this.adminServer.path("/"));
-
-            http.authorizeRequests()
-                .antMatchers(this.adminServer.path("/assets/**")).permitAll() // <1>
-                .antMatchers(this.adminServer.path("/login")).permitAll()
-                .anyRequest().authenticated() // <2>
-                .and()
-            .formLogin().loginPage(this.adminServer.path("/login")).successHandler(successHandler).and() // <3>
-            .logout().logoutUrl(this.adminServer.path("/logout")).and()
-            .httpBasic().and() // <4>
-            .csrf()
-                .csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse()) // <5>
-                .ignoringAntMatchers(
-                    this.adminServer.path("/instances"), // <6>
-                    this.adminServer.path("/actuator/**") // <7>
-                );
-            // @formatter:on
-        }
-    }
-    // end::configuration-spring-security[]
 
     // tag::customization-instance-exchange-filter-function[]
     @Bean
@@ -145,32 +75,4 @@ public class SpringBootAdminServletApplication {
     }
     // end::customization-http-headers-providers[]
 
-
-    // tag::configuration-filtering-notifier[]
-    @Configuration
-    public static class NotifierConfig {
-        private final InstanceRepository repository;
-        private final ObjectProvider<List<Notifier>> otherNotifiers;
-
-        public NotifierConfig(InstanceRepository repository, ObjectProvider<List<Notifier>> otherNotifiers) {
-            this.repository = repository;
-            this.otherNotifiers = otherNotifiers;
-        }
-
-        @Bean
-        public FilteringNotifier filteringNotifier() { // <1>
-            CompositeNotifier delegate = new CompositeNotifier(this.otherNotifiers.getIfAvailable(Collections::emptyList));
-            return new FilteringNotifier(delegate, this.repository);
-        }
-
-        @Primary
-        @Bean(initMethod = "start", destroyMethod = "stop")
-        public RemindingNotifier remindingNotifier() { // <2>
-            RemindingNotifier notifier = new RemindingNotifier(filteringNotifier(), this.repository);
-            notifier.setReminderPeriod(Duration.ofMinutes(10));
-            notifier.setCheckReminderInverval(Duration.ofSeconds(10));
-            return notifier;
-        }
-    }
-    // end::configuration-filtering-notifier[]
 }
